refactor: Move Contains function into test.go
